- Added invalid string sanitize if json serialize fail - Added JSON_SKIP_SANITIZE flag in case it is necessary to disable sanitifcation
