+++ tests/StdClassTest.php
@@ -54,7 +54,7 @@ final class StdClassTest extends TestCase
         $obj->a = 1;
         $obj->b = [1,2,3];
 
-        $serializedValue = JsonSerialize::serialize($obj, JSON_PRETTY_PRINT | JsonSerialize::JSON_SERIALIZE_SKIP_CLASS_NAME);
+        $serializedValue = JsonSerialize::serialize($obj, JSON_PRETTY_PRINT | JsonSerialize::JSON_SKIP_CLASS_NAME);
         $unserializedValue = new stdClass();
         $unserializedValue = JsonSerialize::unserializeToObj($serializedValue, $unserializedValue);
         $this->assertEquals($obj, $unserializedValue, 'Test JSON_SERIALIZE_SKIP_CLASS_NAME flag');
